[maint] Removing hard-coded database prefix

+++ core/components/com_cart/migrations/Migration20180718194712ComCart.php
@@ -3,7 +3,7 @@
 use Hubzero\Content\Migration\Base;
 
 /**
- * Migration script to create an index on jos_cart_meta
+ * Migration script to create an index on `#__cart_meta`
  **/
 class Migration20180718194712ComCart extends Base
 {
@@ -12,9 +12,10 @@ class Migration20180718194712ComCart extends Base
 	 **/
 	public function up()
 	{
-		if (!$this->db->tableHasKey('#__cart_meta', 'idx_scope_id'))
+		if ($this->db->tableExists('#__cart_meta')
+		 && !$this->db->tableHasKey('#__cart_meta', 'idx_scope_id'))
 		{
-			$query = "CREATE INDEX idx_scope_id ON #__cart_meta (scope_id)";
+			$query = "CREATE INDEX idx_scope_id ON `#__cart_meta` (scope_id)";
 			$this->db->setQuery($query);
 			$this->db->query();
 		}
@@ -25,7 +26,8 @@ class Migration20180718194712ComCart extends Base
 	 **/
 	public function down()
 	{
-		if ($this->db->tableHasKey('#__cart_meta', 'idx_scope_id'))
+		if ($this->db->tableExists('#__cart_meta')
+		 && $this->db->tableHasKey('#__cart_meta', 'idx_scope_id'))
 		{
 			$query = "DROP INDEX scope_id_idx ON #__cart_meta";
 			$this->db->setQuery($query);
